+++ lib/loxxy.rb
@@ -2,6 +2,7 @@
 
 require_relative 'loxxy/version'
 require_relative 'loxxy/front_end/raw_parser'
+require_relative 'loxxy/front_end/parser'
 
 module Loxxy
   class Error < StandardError; end